Version 3.29.59 (based on bleeding_edge revision r23884)
Do not use wide reads in CopyCharsUnsigned (Chromium issue 412967).

Fix inaccurate type condition in Hydrogen (Chromium issue 412210).

Fix crash in ScriptDebugServer::wrapCallFrames (Chromium issue 411196).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@23887 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arm64/instructions-arm64.h
@@ -338,7 +338,7 @@ class Instruction {
   // Find the PC offset encoded in this instruction. 'this' may be a branch or
   // a PC-relative addressing instruction.
   // The offset returned is unscaled.
-  ptrdiff_t ImmPCOffset();
+  int64_t ImmPCOffset();
 
   // Find the target of this instruction. 'this' may be a branch or a
   // PC-relative addressing instruction.
